backend/requirements.txt
fastapi==0.111.0
uvicorn==0.24.0
pydantic==2.12.5


APScheduler==3.11.2
python-dotenv==1.0.1

requests==2.32.5
httpx==0.28.1
beautifulsoup4==4.14.3
lxml==6.0.2
tldextract==5.3.1

nltk==3.9.3
textblob==0.17.1
regex==2026.2.28
tqdm==4.67.3
joblib==1.5.3

firebase-admin==6.4.0
google-genai==1.65.0
google-cloud-firestore==2.23.0
google-cloud-storage==3.9.0

protobuf==5.29.6
grpcio==1.78.0

python-dateutil==2.8.2
PyYAML==6.0.3
urllib3==2.6.3
typing_extensions==4.15.0
motor
pymongo
dnspython